+++ monitorcontrol/__init__.py
@@ -4,4 +4,5 @@ from .monitorcontrol import (  # noqa: F401
     get_monitors,
     Monitor,
     PowerMode,
+    InputSource,
 )