+++ monitorcontrol/vcp/vcp_codes.py
@@ -1,3 +1,29 @@
+###############################################################################
+# Copyright 2019-present Alex M.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+###############################################################################
+
+import sys
+
+# f strings require python 3.6
+assert sys.version_info >= (3, 6), "f strings require python 3.6"
 # incomplete list of VCP codes from the MCCS specification
 _VCP_CODE_DEFINTIONS = {
     "image_factory_default": {
@@ -22,6 +48,12 @@ _VCP_CODE_DEFINTIONS = {
         "name": "active control",
         "value": 0x52,
         "type": "ro",
+        "function": "nc",
+    },
+    "input_select": {
+        "name": "input select",
+        "value": 0x60,
+        "type": "rw",
         "function": "nc",
     },
     "image_orientation": {